Refactoring Raven::Event.capture_exception
We would like to be able to subclass Raven::Event to add custom
behavior for things like stack trace sanitization. Before we can do
that, it would be helpful to have smaller functions that we can
override. We have split up Raven::Event.capture_exception to allow
higher granularity in overriding behaviour.

Change-Id: I5dcb35ef1ba3939611472a34f8f5c94b64b835a5
Reviewed-on: https://gerrit.causes.com/9038
Reviewed-by: Aiden Scandella <aiden@causes.com>
Tested-by: Noah Silas <noah@causes.com>